DAOS-3245 control: doc updates and unit tests for storage prepare
