+++ src/control/cmd/dmg/storage_test.go
@@ -89,6 +89,55 @@ func TestStorageCommands(t *testing.T) {
 			cmdSuccess,
 		},
 		{
+			"Prepare without force",
+			"storage prepare",
+			"ConnectClients",
+			nil,
+			fmt.Errorf("consent not given"),
+		},
+		{
+			"Prepare with nvme-only and scm-only",
+			"storage prepare --force --nvme-only --scm-only",
+			"ConnectClients",
+			nil,
+			fmt.Errorf("nvme-only and scm-only options should not be set together"),
+		},
+		{
+			"Prepare with scm-only",
+			"storage prepare --force --scm-only",
+			"ConnectClients StoragePrepare",
+			nil,
+			cmdSuccess,
+		},
+		{
+			"Prepare with nvme-only",
+			"storage prepare --force --nvme-only",
+			"ConnectClients StoragePrepare",
+			nil,
+			cmdSuccess,
+		},
+		{
+			"Prepare with non-existent option",
+			"storage prepare --force --nvme",
+			"",
+			nil,
+			fmt.Errorf("unknown flag `nvme'"),
+		},
+		{
+			"Prepare with force and reset",
+			"storage prepare --force --reset",
+			"ConnectClients StoragePrepare",
+			nil,
+			cmdSuccess,
+		},
+		{
+			"Prepare with force",
+			"storage prepare --force",
+			"ConnectClients StoragePrepare",
+			nil,
+			cmdSuccess,
+		},
+		{
 			"Nonexistent subcommand",
 			"storage quack",
 			"",
