+++ src/control/server/ctl_storage.go
@@ -28,9 +28,39 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/daos-stack/daos/src/control/common"
 	types "github.com/daos-stack/daos/src/control/common/storage"
+	"github.com/daos-stack/daos/src/control/drpc"
+	"github.com/daos-stack/daos/src/control/logging"
 )
+
+// StorageControlService encapsulates the storage part of the control service
+type StorageControlService struct {
+	log  logging.Logger
+	nvme *nvmeStorage
+	scm  *scmStorage
+	drpc drpc.DomainSocketClient
+}
+
+// NewStorageControlService returns an initialized *StorageControlService
+func NewStorageControlService(log logging.Logger, cfg *Configuration) (*StorageControlService, error) {
+	scriptPath, err := cfg.ext.getAbsInstallPath(spdkSetupPath)
+	if err != nil {
+		return nil, err
+	}
+
+	spdkScript := &spdkSetup{
+		log:         log,
+		scriptPath:  scriptPath,
+		nrHugePages: cfg.NrHugepages,
+	}
+
+	return &StorageControlService{
+		log:  log,
+		nvme: newNvmeStorage(log, cfg.NvmeShmID, spdkScript, cfg.ext),
+		scm:  newScmStorage(log, cfg.ext),
+		drpc: getDrpcClientConnection(cfg.SocketDir),
+	}, nil
+}
 
 type PrepareNvmeRequest struct {
 	HugePageCount int
@@ -43,7 +73,7 @@ type PrepareNvmeRequest struct {
 //
 // Suitable for commands invoked directly on server, not over gRPC.
 func (c *StorageControlService) PrepareNvme(req PrepareNvmeRequest) error {
-	ok, usr := common.CheckSudo()
+	ok, usr := c.nvme.ext.checkSudo()
 	if !ok {
 		return errors.Errorf("%s must be run as root or sudo", os.Args[0])
 	}
@@ -79,7 +109,7 @@ type PrepareScmRequest struct {
 //
 // Suitable for commands invoked directly on server, not over gRPC.
 func (c *StorageControlService) PrepareScm(req PrepareScmRequest) (needsReboot bool, pmemDevs []pmemDev, err error) {
-	ok, _ := common.CheckSudo()
+	ok, _ := c.scm.ext.checkSudo()
 	if !ok {
 		err = errors.Errorf("%s must be run as root or sudo", os.Args[0])
 		return
